requirements.txt
playwright==1.54.0
pandas>=2.0
google-api-python-client>=2.70
google-auth>=2.20
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
rambow==1.0.1
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
openpyxl>=3.1.0
